add boxes to primitives node. fixes #2
