+++ examples/01-primitives/index.html
@@ -101,7 +101,9 @@ async function main () {
 
     pNode.filledEllipse([ 192, 140 ], 8, 8, 20, [ 1, 1, 1, 0.9 ])
 
-    pNode.filledBox([ 192, 170 ], 30, 24, [ 0, 1, 0, 1 ])
+    pNode.filledBox([ 50, 170 ], 42, 23, [ 0, 1, 0, 1 ])
+
+    pNode.box([ 120, 130 ], 42, 23, [ 1, 0, 1, 1 ], 1)
 
 
     // use resizeViewport to init values on load:
